temporary2

+++ Shaders/clouds.fs
@@ -7,20 +7,20 @@ out vec4 fragColor;
 // Contact the author for other licensing options
 
 /*
-	Technical details:
+    Technical details:
 
-	The main volume noise is generated from a deformed periodic grid, which can produce
-	a large range of noise-like patterns at very cheap evalutation cost. Allowing for multiple
-	fetches of volume gradient computation for improved lighting.
+    The main volume noise is generated from a deformed periodic grid, which can produce
+    a large range of noise-like patterns at very cheap evalutation cost. Allowing for multiple
+    fetches of volume gradient computation for improved lighting.
 
-	To further accelerate marching, since the volume is smooth, more than half the the density
-	information isn't used to rendering or shading but only as an underlying volume	distance to 
-	determine dynamic step size, by carefully selecting an equation	(polynomial for speed) to 
-	step as a function of overall density (not necessarily rendered) the visual results can be 
-	the	same as a naive implementation with ~40% increase in rendering performance.
+    To further accelerate marching, since the volume is smooth, more than half the the density
+    information isn't used to rendering or shading but only as an underlying volume	distance to 
+    determine dynamic step size, by carefully selecting an equation	(polynomial for speed) to 
+    step as a function of overall density (not necessarily rendered) the visual results can be 
+    the	same as a naive implementation with ~40% increase in rendering performance.
 
-	Since the dynamic marching step size is even less uniform due to steps not being rendered at all
-	the fog is evaluated as the difference of the fog integral at each rendered step.
+    Since the dynamic marching step size is even less uniform due to steps not being rendered at all
+    the fog is evaluated as the difference of the fog integral at each rendered step.
 
 */
 
@@ -34,7 +34,7 @@ vec2 bsMo = vec2(0);
 vec2 disp(float t){ return vec2(sin(t*0.22)*1., cos(t*0.175)*1.)*2.; }
 
 uniform float iTime;
-uniform ivec2 iResolution;
+uniform vec2 iResolution;
 
 
 
@@ -51,7 +51,7 @@ vec2 map(vec3 p)
     float dspAmp = 0.1 + prm1*0.2;
     for(int i = 0; i < 5; i++)
     {
-		p += sin(p.zxy*0.75*trk + iTime*trk*.8)*dspAmp;
+        p += sin(p.zxy*0.75*trk + iTime*trk*.8)*dspAmp;
         d -= abs(dot(cos(p), sin(p.yzx))*z);
         z *= 0.57;
         trk *= 1.4;
@@ -63,39 +63,38 @@ vec2 map(vec3 p)
 
 vec4 render( in vec3 ro, in vec3 rd, float time )
 {
-	vec4 rez = vec4(0);
+    vec4 rez = vec4(0);
     const float ldst = 8.;
-	vec3 lpos = vec3(disp(time + ldst)*0.5, time + ldst);
-	float t = 1.5;
-	float fogT = 0.;
-	for(int i=0; i<130; i++)
-	{
-		if(rez.a > 0.99)break;
+    vec3 lpos = vec3(disp(time + ldst)*0.5, time + ldst);
+    float t = 1.5;
+    float fogT = 0.;
+    for(int i=0; i<130; i++)
+    {
+        if(rez.a > 0.99)break;
 
-		vec3 pos = ro + t*rd;
+        vec3 pos = ro + t*rd;
         vec2 mpv = map(pos);
-		float den = clamp(mpv.x-0.3,0.,1.)*1.12;
-		float dn = clamp((mpv.x + 2.),0.,3.);
+        float den = clamp(mpv.x-0.3,0.,1.)*1.12;
+        float dn = clamp((mpv.x + 2.),0.,3.);
         
-		vec4 col = vec4(0);
+        vec4 col = vec4(0);
         if (mpv.x > 0.6)
-        {
-        
+        {        
             col = vec4(sin(vec3(5.,0.4,0.2) + mpv.y*0.1 +sin(pos.z*0.4)*0.5 + 1.8)*0.5 + 0.5,0.08);
             col *= den*den*den;
-			col.rgb *= linstep(4.,-2.5, mpv.x)*2.3;
+            col.rgb *= linstep(4.,-2.5, mpv.x)*2.3;
             float dif =  clamp((den - map(pos+.8).x)/9., 0.001, 1. );
             dif += clamp((den - map(pos+.35).x)/2.5, 0.001, 1. );
             col.xyz *= den*(vec3(0.005,.045,.075) + 1.5*vec3(0.033,0.07,0.03)*dif);
         }
-		
-		float fogC = exp(t*0.2 - 2.2);
-		col.rgba += vec4(0.06,0.11,0.11, 0.1)*clamp(fogC-fogT, 0., 1.);
-		fogT = fogC;
-		rez = rez + col*(1. - rez.a);
-		t += clamp(0.5 - dn*dn*.05, 0.09, 0.3);
-	}
-	return clamp(rez, 0.0, 1.0);
+
+        float fogC = exp(t*0.2 - 2.2);
+        col.rgba += vec4(0.06,0.11,0.11, 0.1)*clamp(fogC-fogT, 0., 1.);
+        fogT = fogC;
+        rez = rez + col*(1. - rez.a);
+        t += clamp(0.5 - dn*dn*.05, 0.09, 0.3);
+    }
+    return clamp(rez, 0.0, 1.0);
 }
 
 float getsat(vec3 c)
@@ -118,8 +117,8 @@ vec3 iLerp(in vec3 a, in vec3 b, in float x)
 }
 
 void main()
-{	
-	vec2 q = gl_FragCoord .xy/iResolution.xy;
+{
+    vec2 q = gl_FragCoord .xy/iResolution.xy;
     vec2 p = (gl_FragCoord .xy - 0.5*iResolution.xy)/iResolution.y;
     //bsMo = (iMouse.xy - 0.5*iResolution.xy)/iResolution.y;
     bsMo = vec2(0, 0);
@@ -138,11 +137,11 @@ void main()
     vec3 rightdir = normalize(cross(target, vec3(0,1,0)));
     vec3 updir = normalize(cross(rightdir, target));
     rightdir = normalize(cross(updir, target));
-	vec3 rd=normalize((p.x*rightdir + p.y*updir)*1. - target);
+    vec3 rd=normalize((p.x*rightdir + p.y*updir)*1. - target);
     rd.xy *= rot(-disp(time + 3.5).x*0.2 + bsMo.x);
     prm1 = smoothstep(-0.4, 0.4,sin(iTime*0.3));
-	vec4 scn = render(ro, rd, time);
-		
+    vec4 scn = render(ro, rd, time);
+
     vec3 col = scn.rgb;
     col = iLerp(col.bgr, col.rgb, clamp(1.-prm1,0.05,1.));
     
@@ -150,6 +149,6 @@ void main()
 
     col *= pow( 16.0*q.x*q.y*(1.0-q.x)*(1.0-q.y), 0.12)*0.7+0.3; //Vign
     
-	fragColor = vec4( col, 1.0 );
+    fragColor = vec4( col, 1.0 );
     //fragColor = vec4(1.0, 0.0, 0.0, 1.0);
 }
